Reorganize blender logbook files to bit nicer format
