+++ src/components/BlenderLogbook/BlenderLogbook.module.scss
@@ -1,4 +1,4 @@
-@import '../../../variables.scss';
+@import '../../variables.scss';
 
 .form {
   background: $white;
